+++ dispatcher/Cargo.toml
@@ -18,3 +18,4 @@ cheri = ["machine_interface/cheri"]
 mmu = ["machine_interface/mmu"]
 wasm = ["machine_interface/wasm"]
 wasmtime-jit = ["machine_interface/wasmtime-jit"]
+wasmtime-precomp = ["machine_interface/wasmtime-precomp"]
